requirements.txt
requests
virustotal-python
pyOpenSSL